// Kulthia Dasboard/src/app/core/models/ledger.model.ts
export interface LedgerModel {
    Location: string;
    Series: string;
    LedgerDescription : string;
    id: any;
    SlNo: any;
    Alias: String;
    City: String;
    GroupDescription: String;
    Address: String;
    ZipCode: String;
    SubGroup: String;
    LedgerType: String;
    AcType: String;
    AcHolder: String;
    AcNO: String;
    IfscCode: String;

}